+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Future
+
+### Features
+
+- Block and stream checksums are now validated during decompression.
+
 ## 0.0.0
 
 - First release!